Move the gripping-related tests into a separate module
We'll add more gripping tests for
- `grip_at_position`
- `soft_grip`
- `soft_grip_at_position`
- `strong_grip`
- `strong_grip_at_position`

Having them in a dedicated module makes the test base cleaner and easier
to follow.
